Campus Program license clarification
Under <https://docs.github.com/en/education/explore-the-benefits-of-teaching-and-learning-with-github-education/use-github-at-your-educational-institution/about-github-campus-program#github-campus-program-terms-and-conditions> we can read:
"The license is free for one year and will automatically renew for free every 2 years. You may continue on the free license so long as you continue to operate within the terms of the agreement."

This isn't clear what happens after year 1. So either it renews yearly or there is something special about year 2.
The following sentence is making it even less clear by stating its free as long as you operate within the terms of the agreement. So what is this license renewal all about?

So suggest to change this into:
"The license is free for so long as you continue to operate within the terms of the agreement. It will automatically renew for free every year. "